==> student_submissions/projects/luther/dan_hui/04_Merge_1988-1997.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
import re
pd.set_option('display.float_format', lambda x: '%.2f' % x)

### Set Global Ranges and Variables

#years = range(2008,2018)            #Comment in/out year ranges
years = range(1988,1998)

num_budget_files = 6                #total number of budget files to read

### Set Definitions

#clean time from movie data
def clean_time(row):   
    try:
        return datetime.strptime(row, '%B %d, %Y')
    except: pass    

def clean_time_2(row):
    try:
        return datetime.strptime(row, "%Y-%m-%d")
    except: pass    

#create unique ID from movie open date and title
def trunc(row, characters):
    try:
        row = str(row)    #cast to string
        return row[:characters]
    except:
        return row

#create unique ID from movie open date and title
def trunc_end(row, characters):
    try:
        return row[-characters:]
    except:
        return row

#create unique ID from movie open date and title
def time2string(row):
    try:
        return row.strftime('%Y-%m-%d')
    except: pass
    
#create unique ID from movie open date and title
def time2monthyearstring(row):
    try:
        return row.strftime('%Y-%m')
    except: pass
    
#remove (YEAR) formatting
def remove_year(row):
    return row.split(' (')[0]



### Read In Data

#first read the annual list of movies, starting the the first year's set
year_df = pd.read_csv('../04_Data/{}_movies.csv'.format(years[0]), index_col=0)

#then concatenate the remaining 9 years together
for year in years[1:]:
    temp_df = pd.read_csv('../04_Data/{}_movies.csv'.format(year), index_col=0)
    year_df = pd.concat([year_df,temp_df])
    
#Clean Datetime
year_df["open_date"] = year_df["open_date"].apply(clean_time_2)

#read the movie detail data, starting with the first year's set
movie_df = pd.read_csv('../04_Data/{}_movies_detail.csv'.format(years[0]), index_col=0)

#then concatenate the remaining 9 years together
for year in years[1:]:
    temp_df = pd.read_csv('../04_Data/{}_movies_detail.csv'.format(year), index_col=0)
    movie_df = pd.concat([movie_df, temp_df])
    
#Clean date time from movie dataframe
movie_df["new_close"] = movie_df["close"].apply(clean_time)

#Read in the budget data
budget_df = pd.read_csv('../04_Data/budget_1.csv', index_col=0, parse_dates=['open_date'])

#concatenate remaining data together
for k in range(2,(num_budget_files+1)):                              #Use variable to set # of files to read
    temp_df = pd.read_csv('../04_Data/budget_{}.csv'.format(k), index_col=0, parse_dates=['open_date'])
    budget_df = pd.concat([budget_df, temp_df])

#isolate useful data
budget_df = budget_df.drop(["rank","worldwide"],axis=1)   #drop unecessary columns
budget_df = budget_df[budget_df['open_date'] >= datetime(years[0], 1, 1,0,0)]    #only incl. after the start year
budget_df = budget_df[budget_df['open_date'] <= datetime(years[-1], 12, 31,0,0)] #only incl. before the end of last year

budget_df = budget_df[budget_df['domestic'] > 10000]     #only include movies that grossed at least $10K

budget_df = budget_df.rename(index=str, columns={'open_date':'open'})

### Inspect Data

#Yearly Movie Data
year_df.head(5)

year_df.shape

#Movie Detail
movie_df.head(5)

movie_df.shape

#Budget
budget_df.head(5)

budget_df.shape

### Merge Movies by Year and Movie Tables Together

#Merge the movie year + movie details lists together using the unique URL
detail_df = pd.merge(year_df, movie_df, on='url', how='left')

#re-organize tables to be a little cleaner and leave out uneeded columns
detail_df = detail_df[["title",'mpaa',"open_date","total_box",'open_box',
                     'worldwide gross','genre','runtime','rundays','3d', 'imax', 'series', 'max_sites']]

#rename columns for readability
detail_df = detail_df.rename(index=str, columns={"open_date": "open", "total_box": 'usa_box', 
                              "open_box": 'usa_open', "worldwide gross": "intl_box", "max_sites":'theaters' })

#take a look
detail_df.head(2)

### Check for Duplicates

#check to see if there are any movie table duplicates. There are not. 
temp_df = detail_df.groupby(["title"])['mpaa'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['mpaa'],ascending=False).head(2)

#check to see if there are any budget table duplicates on title and open. There are not. 
temp_df = budget_df.groupby(["title","open"])['prod_budget'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['prod_budget'],ascending=False).head(2)

### Merge Movie and Budget Tables

The Movie table is from BoxOfficeMojo.com and the Budget table is from The-Numbers.com, and there are inconsistencies in how the movies are named. For example, the same movie is called 'Star Wars: The Last Jedi on BoxOfficeMojo and 'Star Wars Ep. VII: The Last Jedi' on The-Numbers. "Tyler Perry's Meet the Browns" on BoxOfficeMojo is "Meet the Browns" on The-Numbers. Also, "Under the Same Moon" on BoxOfficeMojo is "La misma luna" on The-Numbers

Due to the variety of differences in the data, I will try a range of strategies to create unique matches from one table to the others. 

__Merge 1: Unique ID Using Release Date + Forward Title Fragment__

#Create Unique ID from the opening day + the first 5 characters of the title. This is a preferred 
#route over matching to just the movie title because movies are named differently on BoxOfficeMojo vs The Numbers
budget_df["id"] = budget_df["open"].apply(time2monthyearstring) + budget_df["title"].apply(trunc,characters=9)
detail_df["id"] = detail_df["open"].apply(time2monthyearstring) + detail_df["title"].apply(trunc,characters=9)

#check to see if the budget keys are unique. They are!  
temp_df = budget_df.groupby(["id"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

#check to see if the movie keys are unique. They are!  
temp_df = detail_df.groupby(["id"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

#Merge!! 
temp_df = pd.merge(detail_df, budget_df, on='id', how='outer')

#This contains the movies that matched to a budget via 'ID' as a unique ID
group_1_df = temp_df[temp_df["prod_budget"].notnull() & temp_df["title_x"].notnull()]  
group_1_df = group_1_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
group_1_df = group_1_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters","prod_budget"]]

group_1_df.head(2)

group_1_df.shape  #477 Movies Matched on this merge

#this is a dataframe of movies that didn't match
remaining_df = temp_df[temp_df["prod_budget"].isnull()]
remaining_df = remaining_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
remaining_df = remaining_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters"]]

remaining_df.shape #1523 Movies did not Match on this merge

#this is a dataframe of budget information that has not matched yet
remain_budget_df = temp_df[temp_df["title_x"].isnull()]  
remain_budget_df = remain_budget_df.rename(index=str, columns={'title_y':'title','open_y':'open'})
remain_budget_df = remain_budget_df[["open","title","prod_budget","domestic"]]

remain_budget_df.shape  #70 Budget Lines remain

__Merge 2: Unique ID Using Release Date + Trailing Title Fragment__

#create new 'ID2' to match using opening date plus trailing characters from the movie name
remain_budget_df["id2"] = remain_budget_df["open"].apply(time2string)  + remain_budget_df["title"].apply(trunc_end,characters=9)
remaining_df["id2"] = remaining_df["open"].apply(time2string)  + remaining_df["title"].apply(trunc_end,characters=9)

#check to see if the budget keys are unique. They Are!  
temp_df = remain_budget_df.groupby(["id2"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

#check to see if the movie keys are unique. There are two duplicates, which have closer inspection below
temp_df = remaining_df.groupby(["id2"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

# inpsecting the duplicate keys above shows two are re-releases of previous movies and may not have
# had theatrical releases It's probably ok to leave these in there since they would be removed later
# on for not having theaters
remaining_df[(remaining_df['open']=='2012-07-19')]

#Merge!!!!!
temp_df = pd.merge(remaining_df, remain_budget_df, on='id2', how='outer')

#This contains the movies that matched to a budget via 'ID2' as a unique ID
group_2_df = temp_df[temp_df["prod_budget"].notnull() & temp_df["title_x"].notnull()]  
group_2_df = group_2_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
group_2_df = group_2_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters","prod_budget"]]

group_2_df.shape    #9 Movie Matched in this round

#this is a dataframe of movies that didn't match
remaining_2_df = temp_df[temp_df["prod_budget"].isnull()]
remaining_2_df = remaining_2_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
remaining_2_df = remaining_2_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters"]]

remaining_2_df.shape  #1514 Movies remain unmatched 

#this is a dataframe of budget information that has not matched yet
remain_budget_2_df = temp_df[temp_df["title_x"].isnull()]  
remain_budget_2_df = remain_budget_2_df.rename(index=str, columns={'title_y':'title','open_y':'open'})
remain_budget_2_df = remain_budget_2_df[["open","title","prod_budget","domestic"]]

remain_budget_2_df.shape   #61 Budget Lines remain

__Merge 3: Titles__

This will catch movies that have different open dates on BoxOfficeMojo vs The-Numbers

#Some titles in Box Office Mojo have Years in their names with (). These should be cleaned
remaining_2_df["title"] = remaining_2_df["title"].apply(remove_year)

#Since the budget rows are only unique on title and also date, I'll still use year and month
remaining_2_df["id"] = remaining_2_df["open"].apply(time2monthyearstring) + remaining_2_df["title"]
remain_budget_2_df["id"] = remain_budget_2_df["open"].apply(time2monthyearstring) + remain_budget_2_df["title"]

#merge on the name of the movie itself
temp_df = pd.merge(remaining_2_df, remain_budget_2_df, on='id', how='outer')

#this contains the movies that matched by name
group_3_df = temp_df[temp_df["prod_budget"].notnull() & temp_df["mpaa"].notnull()]    #boolean changed to mpaa here
group_3_df = group_3_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
group_3_df = group_3_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters","prod_budget"]]

group_3_df.shape  #2 Movies matched on Title

#check to make sure there were no duplicates. Looks good! 
group_3_df.groupby(["title"])["mpaa"].count().reset_index().sort_values(["mpaa"],ascending=False).head(2)

#this is a dataframe of movies that didn't match
remaining_3_df = temp_df[temp_df["prod_budget"].isnull()]
remaining_3_df = remaining_3_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
remaining_3_df = remaining_3_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters"]]

remaining_3_df.shape  #1512 Movies remain unmatched

#this is a dataframe of budget information that has not matched yet
remain_budget_3_df = temp_df[temp_df["mpaa"].isnull()]    #boolean changed to mpaa here
remain_budget_3_df = remain_budget_3_df.rename(index=str, columns={'title_y':'title','open_y':'open'})
remain_budget_3_df = remain_budget_3_df[["open","title","prod_budget","domestic"]]

remain_budget_3_df.shape  #59 Budgets Remain Unmatched

__Explore Remaining Titles: What Else Can I Merge On?__

remaining_3_df.sort_values(['usa_box'],ascending=False).head(5)

remain_budget_3_df.sort_values(['domestic'],ascending=False).head(5)

I also think I can match on Open Date and the first few digits of the domestic box office

__Merge 4: Date + First N Digits of the Domestic Box Office Total__

#Create Unique ID from the opening day + the first 2 characters of the domestic budget. 
remain_budget_3_df["id"] = remain_budget_3_df["open"].apply(time2monthyearstring)  + '-' + remain_budget_3_df["domestic"].apply(trunc,characters=3)
remaining_3_df["id"] = remaining_3_df["open"].apply(time2monthyearstring)  + '-' + remaining_3_df["usa_box"].apply(trunc,characters=3)

#check to see if the budget keys are unique. They are!  
temp_df = remain_budget_3_df.groupby(["id"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

#merge!!!
temp_df = pd.merge(remaining_3_df, remain_budget_3_df, on='id', how='outer')

#this contains the movies that matched by name
group_4_df = temp_df[temp_df["prod_budget"].notnull() & temp_df["title_x"].notnull()]    
group_4_df = group_4_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
group_4_df = group_4_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters","prod_budget"]]

group_4_df.shape  #15 Movies Matched

#this is a dataframe of movies that didn't match
remaining_4_df = temp_df[temp_df["prod_budget"].isnull()]
remaining_4_df = remaining_4_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
remaining_4_df = remaining_4_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters"]]

remaining_4_df.shape   #1497 Movies Remail Unmatched

#this is a dataframe of budget information that has not matched yet
remain_budget_4_df = temp_df[temp_df["mpaa"].isnull()]    #boolean changed to mpaa here
remain_budget_4_df = remain_budget_4_df.rename(index=str, columns={'title_y':'title','open_y':'open'})
remain_budget_4_df = remain_budget_4_df[["open","title","prod_budget","domestic"]]

remain_budget_4_df.shape   #44 Budgets unmatched

__Merge 5: Title + First N Digits of the Domestic Box Office Total__

#Create Unique ID from the opening day + the first 2 characters of the domestic budget. 
remain_budget_4_df["id"] = remain_budget_4_df["title"] + remain_budget_4_df["domestic"].apply(trunc,characters=2)
remaining_4_df["id"] = remaining_4_df["title"] + remaining_4_df["usa_box"].apply(trunc,characters=2)

#check to see if the budget keys are unique. They are!  
temp_df = remain_budget_4_df.groupby(["id"])['title'].count()
temp_df = temp_df.reset_index()
temp_df.sort_values(['title'],ascending=False).head(2)

#merge!!!
temp_df = pd.merge(remaining_4_df, remain_budget_4_df, on='id', how='outer')

#this contains the movies that matched by name
group_5_df = temp_df[temp_df["prod_budget"].notnull() & temp_df["title_x"].notnull()]    
group_5_df = group_5_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
group_5_df = group_5_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters","prod_budget"]]

group_5_df.shape      #5 Movie Matched this time

#this is a dataframe of movies that didn't match
remaining_5_df = temp_df[temp_df["prod_budget"].isnull()]
remaining_5_df = remaining_5_df.rename(index=str, columns={'title_x':'title','open_x':'open'})
remaining_5_df = remaining_5_df[["title","mpaa","open","usa_box","usa_open","intl_box","genre","runtime",
                         "rundays","3d","imax","series","theaters"]]

remaining_5_df.shape   #1492 Remain Unmatched

#this is a dataframe of budget information that has not matched yet
remain_budget_5_df = temp_df[temp_df["mpaa"].isnull()]    #boolean changed to mpaa here
remain_budget_5_df = remain_budget_5_df.rename(index=str, columns={'title_y':'title','open_y':'open'})
remain_budget_5_df = remain_budget_5_df[["open","title","prod_budget","domestic"]]

remain_budget_5_df.shape  #39 budgets remain unmatched

__Explore Remaining Titles: Nothing Left to Merge On__

I feel pretty good stopping here. The largest movie on the remaining budget dataframe
has a domestic gross of 30 million while the top movie on the remaining movies is $90M which means all the movies in between will never be matched to the budget information, and it is thus not available. That's not bad. Also, there are movies on this list that aren't really from this decade, such as The Lion King, Monsters, Inc, so many of these unmatched movies are rereleases and shouldn't impact a prediction model for this decade's films

remaining_5_df.sort_values(["usa_box"],ascending=False).head(10)

remain_budget_5_df.sort_values(["domestic"],ascending=False).head(10)

__Concatenate & Export the Merged List__

pd.concat([group_1_df, group_2_df, group_3_df, 
            group_4_df, group_5_df, remaining_5_df]).to_csv('../04_Data/{}_{}_merged.csv'.format(years[0],years[-1]))

pd.concat([group_1_df, group_2_df, group_3_df, 
           group_4_df, group_5_df, remaining_5_df]).shape
