==> student_submissions/projects/luther/dan_hui/06_Featurize.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
import seaborn as sns
import re
import matplotlib.pyplot as plt
pd.set_option('display.float_format', lambda x: '%.2f' % x)
import matplotlib.pyplot as plt
from IPython.display import display
pd.options.display.max_columns = None

#years = [2008,2017]   #Comment in/out year ranges
years = [1988,1997] 
movie_df = pd.read_csv('../04_Data/{}_{}_merged.csv'.format(years[0],years[-1]), index_col=0)

movie_df.head()

def find_genre(row, genre):
    try: 
        if re.search(genre, row):
            return 1
        else: return 0
    except: 0 

movie_df["genre"].value_counts().head(25)

#Create Columns for each Unique Genre Category in the top 25
movie_df["action"] = movie_df["genre"].apply(find_genre, genre="Action")        #1 
movie_df["adventure"] = movie_df["genre"].apply(find_genre, genre="Adventure")  #X (combine with Action)
movie_df["comedy"] = movie_df["genre"].apply(find_genre, genre="Comedy")        #2
movie_df["drama"] = movie_df["genre"].apply(find_genre, genre="Drama")          #3
movie_df["horror"] = movie_df["genre"].apply(find_genre, genre="Horror")        #4
movie_df["musical"] = movie_df["genre"].apply(find_genre, genre="Musical")      #X (combine with Comedy)
movie_df["scifi"] = movie_df["genre"].apply(find_genre, genre="Sci")            #5
movie_df["romance"] = movie_df["genre"].apply(find_genre, genre="Roman")        #6
movie_df["animation"] = movie_df["genre"].apply(find_genre, genre="Animat")     #7
#movie_df["foreign"] = movie_df["genre"].apply(find_genre, genre="Foreign")     #X (withdrawn - too niche)
movie_df["thriller"] = movie_df["genre"].apply(find_genre, genre="Thrill")      #X (combine with Action)
#movie_df["documentary"] = movie_df["genre"].apply(find_genre, genre="Document")#X (withdrawn - too niche)
movie_df["fantasy"] = movie_df["genre"].apply(find_genre, genre="Fantasy")      #X (combine with SciFi)

#Compress categories. 
movie_df["comedy"] = (movie_df["comedy"] + movie_df["musical"]) #based on Golden Globes Category Comedy/Musical
#these categories are really similar
movie_df["action"] = (movie_df["action"] + movie_df["adventure"] + movie_df["thriller"])
movie_df["scifi/fantasy"] = movie_df["scifi"] + movie_df["fantasy"]

#make sure each compressed category is max value 1
movie_df["comedy"] = movie_df["comedy"].apply(lambda x: min(1, x))
movie_df["action"] = movie_df["action"].apply(lambda x: min(1, x))
movie_df["scifi/fantasy"] = movie_df["scifi/fantasy"].apply(lambda x: min(1, x))

movie_df.drop(["musical","adventure","fantasy","thriller","scifi"],axis=1,inplace=True)

#Create Columns To Sum the Number of Genres Present as a way to find the Zeros
movie_df["genre_counts"] =  (movie_df["action"] + movie_df["comedy"] +
                             movie_df["drama"] + movie_df["horror"] + movie_df["scifi/fantasy"] +
                             movie_df["romance"] + movie_df["animation"] )

#Create a Genre for 'Other'
movie_df["other"] = movie_df["genre_counts"].apply(lambda x: 1 if x == 0 else 0)

#create unique ID from movie open date and title
def time2monthyearstring(row):
    try:
        return row[0:4]
    except: pass

#truncate the year
movie_df["year"] = movie_df["open"].apply(time2monthyearstring)

#this resets intdex before dummifying the MPAA rating
movie_df = movie_df.reset_index()      

#this creates dummy variables for the MPAA rating
movie_df = pd.merge(movie_df, pd.get_dummies(movie_df["mpaa"]),left_index=True, right_index=True)

movie_df.head(3)

movie_df.to_csv('../04_Data/{}_{}_merged_featurized.csv'.format(years[0],years[-1]))
